pyproject.toml
[project]
name = "adaptive-traffic-signal-green-corridor"
version = "0.1.0"
description = "Adaptive traffic signal control with green corridor and emergency preemption"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python-headless>=4.10",
    "pillow>=9.0",
    "ultralytics>=8.3",
    "pyserial>=3.5",
    "flask>=3.0",
    "numpy>=2.0",
    "pyyaml>=6.0",
    "kaggle>=2.0.0",
    "tk>=0.1.0",
]

[project.optional-dependencies]
prediction = [
    # Optional LSTM / XGBoost predictor (Week 4)
    "scikit-learn>=1.5",
    "xgboost>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["control", "detection", "gui"]


[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
